refactor(exceptions): rename module `traps` to `exceptions`
Trapping is the act of handling exceptions, but this module does not
deal with that aspect.
