+++ src/riscv/lib/src/interpreter/rv64d.rs
@@ -13,6 +13,7 @@ use rustc_apfloat::ieee::Double;
 use rustc_apfloat::ieee::Single;
 
 use super::float::FloatExt;
+use crate::exceptions::Exception;
 use crate::machine_state::MachineCoreState;
 use crate::machine_state::hart_state::HartState;
 use crate::machine_state::memory;
@@ -21,7 +22,6 @@ use crate::machine_state::registers::FValue;
 use crate::machine_state::registers::XRegister;
 use crate::parser::instruction::InstrRoundingMode;
 use crate::state_backend as backend;
-use crate::traps::Exception;
 
 impl From<Double> for FValue {
     fn from(f: Double) -> Self {
@@ -371,6 +371,7 @@ mod tests {
     use proptest::prelude::*;
 
     use crate::backend_test;
+    use crate::exceptions::Exception;
     use crate::machine_state::MachineCoreState;
     use crate::machine_state::hart_state::HartState;
     use crate::machine_state::memory::M4K;
@@ -380,7 +381,6 @@ mod tests {
     use crate::machine_state::registers::parse_xregister;
     use crate::machine_state::registers::t0;
     use crate::state::NewState;
-    use crate::traps::Exception;
 
     backend_test!(test_fmv_d, F, {
         let state = HartState::<F>::new();
