+++ src/riscv/lib/src/lib.rs
@@ -7,6 +7,7 @@ mod array_utils;
 mod bits;
 mod cache_utils;
 mod default;
+mod exceptions;
 mod instruction_context;
 mod interpreter;
 pub mod jit;
@@ -22,4 +23,3 @@ pub mod state_backend;
 mod state_context;
 pub mod stepper;
 pub mod storage;
-mod traps;
